Add function OVERLOADING

+++ src/main/kotlin/Rectangle.kt
@@ -1,4 +1,4 @@
-class Rectangle(val height: Int, val width: Int) : Shape("Rectangle") {
+class Rectangle(val height: Double, val width: Double) : Shape("Rectangle") {
     init {
         // I can use the name property because it's inherited from the parent class
         println("The height of the $name is $height and the width is $width")
@@ -6,7 +6,7 @@ class Rectangle(val height: Int, val width: Int) : Shape("Rectangle") {
 
     //    Secondary constructor
     //for example we want to create a square
-    constructor(side: Int) : this(side, side) {
+    constructor(side: Double) : this(side, side) {
         println("The height of the $name is $height and the width is $width")
     }
     //we can create multiple secondary constructors, the purpose is to have multiple ways to create an object
@@ -17,12 +17,12 @@ class Rectangle(val height: Int, val width: Int) : Shape("Rectangle") {
         return name
     }
 
-    override fun calculateArea(): Int {
+    override fun calculateArea(): Double {
         // override means we are overriding the parent function
         return height * width
     }
 
-    override fun calculatePerimeter(): Int {
+    override fun calculatePerimeter(): Double {
         // we must override this function because it's abstract in the parent class
         return 2 * (height + width)
     }
